frontend/src/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
mysqlclient==2.1.0
mysql-connector-python==8.0.30